uscan: implement 'gitexport' option
The file exclusions are now only overridden if the new option
'gitexport' is set to 'all' on the watch line. Otherwise, the old
behavior of respecting upstream's export rules is retained.

+++ lib/Devscripts/Uscan/Downloader.pm
@@ -51,6 +51,9 @@ has destdir => (is => 'rw');
 has gitrepo_state => (
     is      => 'rw',
     default => sub { 0 });
+has git_export_all => (
+    is      => 'rw',
+    default => sub { 0 });
 has user_agent => (
     is      => 'rw',
     lazy    => 1,
@@ -128,44 +131,49 @@ sub download ($$$$$$$$) {
             uscan_exec('tar', '-C', $tempdir, '-cvf',
                 "$abs_dst/$pkg-$ver.tar", "$pkg-$ver");
         } elsif ($self->git_upstream) {
-            # override any export-subst and export-ignore attributes
             my ($infodir, $attr_file, $attr_bkp);
-            spawn(
-                exec      => [qw|git rev-parse --git-path info/|],
-                to_string => \$infodir,
-            );
-            chomp $infodir;
-            mkdir $infodir unless -e $infodir;
-            spawn(
-                exec      => [qw|git rev-parse --git-path info/attributes|],
-                to_string => \$attr_file,
-            );
-            chomp $attr_file;
-            spawn(
-                exec => [qw|git rev-parse --git-path info/attributes-uscan|],
-                to_string => \$attr_bkp,
-            );
-            chomp $attr_bkp;
-            rename $attr_file, $attr_bkp if -e $attr_file;
-            my $attr_fh;
+            if ($self->git_export_all) {
+                # override any export-subst and export-ignore attributes
+                spawn(
+                    exec      => [qw|git rev-parse --git-path info/|],
+                    to_string => \$infodir,
+                );
+                chomp $infodir;
+                mkdir $infodir unless -e $infodir;
+                spawn(
+                    exec => [qw|git rev-parse --git-path info/attributes|],
+                    to_string => \$attr_file,
+                );
+                chomp $attr_file;
+                spawn(
+                    exec =>
+                      [qw|git rev-parse --git-path info/attributes-uscan|],
+                    to_string => \$attr_bkp,
+                );
+                chomp $attr_bkp;
+                rename $attr_file, $attr_bkp if -e $attr_file;
+                my $attr_fh;
 
-            unless (open($attr_fh, '>', $attr_file)) {
-                rename $attr_bkp, $attr_file if -e $attr_bkp;
-                uscan_die("could not open $attr_file for writing");
+                unless (open($attr_fh, '>', $attr_file)) {
+                    rename $attr_bkp, $attr_file if -e $attr_bkp;
+                    uscan_die("could not open $attr_file for writing");
+                }
+                print $attr_fh "* -export-subst\n* -export-ignore\n";
+                close $attr_fh;
             }
-            print $attr_fh "* -export-subst\n* -export-ignore\n";
-            close $attr_fh;
 
             uscan_exec_no_fail('git', 'archive', '--format=tar',
                 "--prefix=$pkg-$ver/", "--output=$abs_dst/$pkg-$ver.tar",
                 $gitref) == 0
               or uscan_die("git archive failed");
 
-            # restore attributes
-            if (-e $attr_bkp) {
-                rename $attr_bkp, $attr_file;
-            } else {
-                unlink $attr_file;
+            if ($self->git_export_all) {
+                # restore attributes
+                if (-e $attr_bkp) {
+                    rename $attr_bkp, $attr_file;
+                } else {
+                    unlink $attr_file;
+                }
             }
         } else {
             if ($self->gitrepo_state == 0) {
@@ -181,31 +189,33 @@ sub download ($$$$$$$$) {
                     $self->gitrepo_state(2);
                 }
             }
-            # override any export-subst and export-ignore attributes
-            my ($infodir, $attr_file);
-            spawn(
-                exec => [
-                    'git', "--git-dir=$destdir/$gitrepo_dir",
-                    'rev-parse', '--git-path', 'info/'
-                ],
-                to_string => \$infodir,
-            );
-            chomp $infodir;
-            mkdir $infodir unless -e $infodir;
-            spawn(
-                exec => [
-                    'git',       "--git-dir=$destdir/$gitrepo_dir",
-                    'rev-parse', '--git-path',
-                    'info/attributes'
-                ],
-                to_string => \$attr_file,
-            );
-            chomp $attr_file;
-            my $attr_fh;
-            uscan_die("could not open $attr_file for writing")
-              unless open($attr_fh, '>', $attr_file);
-            print $attr_fh "* -export-subst\n* -export-ignore\n";
-            close $attr_fh;
+            if ($self->git_export_all) {
+                # override any export-subst and export-ignore attributes
+                my ($infodir, $attr_file);
+                spawn(
+                    exec => [
+                        'git', "--git-dir=$destdir/$gitrepo_dir",
+                        'rev-parse', '--git-path', 'info/'
+                    ],
+                    to_string => \$infodir,
+                );
+                chomp $infodir;
+                mkdir $infodir unless -e $infodir;
+                spawn(
+                    exec => [
+                        'git',       "--git-dir=$destdir/$gitrepo_dir",
+                        'rev-parse', '--git-path',
+                        'info/attributes'
+                    ],
+                    to_string => \$attr_file,
+                );
+                chomp $attr_file;
+                my $attr_fh;
+                uscan_die("could not open $attr_file for writing")
+                  unless open($attr_fh, '>', $attr_file);
+                print $attr_fh "* -export-subst\n* -export-ignore\n";
+                close $attr_fh;
+            }
 
             uscan_exec_no_fail(
                 'git',                 "--git-dir=$destdir/$gitrepo_dir",
